requirements.txt
asgiref==3.6.0
Django==3.2.13
django-background-tasks==1.2.5
django-compat==1.0.15
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
